Update utils.py
There is a problem running it on Windows. 
result = re.search(split_regex_pattern, file_path)
*** re.error: bad escape \P at position *
When it is reading the string of the 'split_regex_pattern' in the function _parse line 493 of the library sre_parse.py the variable 'split_regex_pattern is an string and take this format: 'data_csv\Prueba_Span\data_(?P<span>.*)\[^/]*'
Then when this variable is been read, it takes the string '\P' as a regular expresion syntax instead of taking it like this '\\'. Thats why we need it to recplace each '\' by '\\'